Make inspectorplugin build again
The Qt Designer plugin has not compiled for a while. It is excluded
from all builds and CI never touches it, so nothing noticed. Three
things were wrong, and each hid the next:

- inspectorplugin.h included libmscore/mscore.h and then declared its
  own stand-in MScore, which is a plain redefinition. The stand-ins are
  deliberate: Designer loads this plugin on its own, so it links
  neither libmscore nor mscoreapp, and instead supplies the few symbols
  the files it compiles expect. Move them into the .cpp, where they do
  not collide, and where mscoreapp's header glob cannot reach them.

- icons.cpp reaches libmscore/sym.h, which includes ft2build.h, so the
  plugin needs freetype's include directories. Link freetype for them.

- On APPLE the plugin's sources were excluded from AUTOGEN, which left
  its Q_OBJECT classes without a vtable, so it could never have linked.
  That exclusion was really working around mscoreapp globbing every
  header in this tree, inspectorplugin.h included, and generating moc
  code for plugin classes it has no implementation of. Drop that header
  from the glob instead, and let moc run for the plugin.

miconengine.cpp also wanted Preferences::getColor, which the stand-in
did not have.
